UnDRY the split by separator code
I should have reflected more when the compiler told me there was dead code if the Markdown feature wasn't enabled around the `SemanticSplitPosition`.

I was trying to overfit too much structure in order to reuse one loop. This was already vastly different between text and markdown, and made the code harder to reason about. Just split the two implementations, and also got rid of the `Level` trait which I also didn't feel great about because the methods there were only ever really needed for Markdown anyway.
