+++ src/lib.rs
@@ -3,12 +3,9 @@
 
 use std::{
     cmp::Ordering,
-    fmt,
-    iter::once,
     ops::{Range, RangeFrom, RangeFull, RangeInclusive, RangeTo, RangeToInclusive},
 };
 
-use either::Either;
 use itertools::Itertools;
 
 mod characters;
@@ -198,28 +195,6 @@ impl ChunkCapacity for RangeToInclusive<usize> {
     }
 }
 
-/// How a particular semantic level relates to surrounding text elements.
-#[allow(dead_code)]
-#[derive(Copy, Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
-enum SemanticSplitPosition {
-    /// The semantic level should be included in the previous chunk.
-    Prev,
-    /// The semantic level should be treated as its own chunk.
-    Own,
-    /// The semantic level should be included in the next chunk.
-    Next,
-}
-
-/// Information required by generic Semantic Levels
-trait Level: fmt::Debug {
-    fn split_position(&self) -> SemanticSplitPosition;
-
-    /// Whether or not when splitting ranges, whitespace should be included as previous.
-    fn treat_whitespace_as_previous(&self) -> bool {
-        false
-    }
-}
-
 /// Implementation that dictates the semantic split points available.
 /// For plain text, this goes from characters, to grapheme clusters, to words,
 /// to sentences, to linebreaks.
@@ -227,7 +202,7 @@ trait Level: fmt::Debug {
 /// lists, and code blocks.
 trait SemanticSplit {
     /// Internal type used to represent the level of semantic splitting.
-    type Level: Copy + Level + Ord + PartialOrd + 'static;
+    type Level: Copy + Ord + PartialOrd + 'static;
 
     /// Levels that are always considered in splitting text, because they are always present.
     const PERSISTENT_LEVELS: &'static [Self::Level];
@@ -522,82 +497,6 @@ where
             }
         }
     }
-}
-
-/// Given a list of separator ranges, construct the sections of the text
-fn split_str_by_separator<L: Level>(
-    text: &str,
-    separator_ranges: impl Iterator<Item = (L, Range<usize>)>,
-) -> impl Iterator<Item = (usize, &str)> {
-    let mut cursor = 0;
-    let mut final_match = false;
-    separator_ranges
-        .batching(move |it| {
-            loop {
-                match it.next() {
-                    // If we've hit the end, actually return None
-                    None if final_match => return None,
-                    // First time we hit None, return the final section of the text
-                    None => {
-                        final_match = true;
-                        return text.get(cursor..).map(|t| Either::Left(once((cursor, t))));
-                    }
-                    // Return text preceding match + the match
-                    Some((level, range)) => {
-                        let offset = cursor;
-                        match level.split_position() {
-                            SemanticSplitPosition::Prev => {
-                                if range.end < cursor {
-                                    continue;
-                                }
-                                let section = text
-                                    .get(cursor..range.end)
-                                    .expect("invalid character sequence");
-                                cursor = range.end;
-                                return Some(Either::Left(once((offset, section))));
-                            }
-                            SemanticSplitPosition::Own => {
-                                if range.start < cursor {
-                                    continue;
-                                }
-                                let prev_section = text
-                                    .get(cursor..range.start)
-                                    .expect("invalid character sequence");
-                                if prev_section.trim().is_empty()
-                                    && level.treat_whitespace_as_previous()
-                                {
-                                    let section = text
-                                        .get(cursor..range.end)
-                                        .expect("invalid character sequence");
-                                    cursor = range.end;
-                                    return Some(Either::Left(once((offset, section))));
-                                }
-                                let separator = text
-                                    .get(range.start..range.end)
-                                    .expect("invalid character sequence");
-                                cursor = range.end;
-                                return Some(Either::Right(
-                                    [(offset, prev_section), (range.start, separator)].into_iter(),
-                                ));
-                            }
-                            SemanticSplitPosition::Next => {
-                                if range.start < cursor {
-                                    continue;
-                                }
-                                let prev_section = text
-                                    .get(cursor..range.start)
-                                    .expect("invalid character sequence");
-                                // Separator will be part of the next chunk
-                                cursor = range.start;
-                                return Some(Either::Left(once((offset, prev_section))));
-                            }
-                        }
-                    }
-                }
-            }
-        })
-        .flatten()
-        .filter(|(_, s)| !s.is_empty())
 }
 
 #[cfg(test)]
